requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mira_sdk==0.1.13
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
plotly==5.24.1
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
scapy==2.6.0
selenium==4.28.0
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.0.0
tomli==2.2.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
